+++ include/RiskManager.hpp
@@ -13,6 +13,6 @@ public:
 
 private:
   std::shared_ptr<PositionManager> position_manager_;
-  const int32_t max_position_per_symbol_ = 1000;
+  const int64_t max_position_per_symbol_ = 1000;
   const double max_order_value_ = 10000.00;
 };
